+++ src/RequestHandler.php
@@ -12,11 +12,11 @@ declare(strict_types=1);
 
 namespace WeCodeIn\Http\Server;
 
-use Interop\Http\Factory\ResponseFactoryInterface;
-use Interop\Http\Server\RequestHandlerInterface;
-use Interop\Http\Server\MiddlewareInterface;
+use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 
 class RequestHandler implements RequestHandlerInterface
 {
@@ -29,7 +29,7 @@ class RequestHandler implements RequestHandlerInterface
         $this->middlewares = $middlewares;
     }
 
-    public function handle(ServerRequestInterface $request) : ResponseInterface
+    public function handle(ServerRequestInterface $request): ResponseInterface
     {
         $handler = clone $this;